Add accuracy evaluation

+++ language/llama2-70b/SUT.py
@@ -262,7 +262,7 @@ class SUTServer(SUT):
 
             output_tokens = tokens_streamer.get_out_tokens()
 
-            response_array = array.array("B", np.array(output_tokens, np.float32).tobytes())
+            response_array = array.array("B", np.array(output_tokens, np.int32).tobytes())
             bi = response_array.buffer_info()
             response = [lg.QuerySampleResponse(
                 qitem.id, bi[0], bi[1])]
